// types.ts
export interface FormValues {
    name: string;
    email: string;
    phone: string;
    address: string;
  };
   
export interface FormValuesService {
    TypeofVehicle: string;
    VehiclemodelNumber: string;
  };
export interface FormValuesBooking {
    appointmentDate: string;
    appointmentTime: string;
  };